무기 차징 로직 변경
차징 게이지 증가 시간 텀 줄여서 게이지 자잘자잘하게 오르게 해야겠음.
총알 코스트 소모 개선할 예정인데 일단 지금 총알 코스트 소모 안됨.
ingame UI 작업좀 해야되는데...

+++ Assets/ItemAsset/Scripts/ItemUseEffect/CharacterTargetEffect.cs
@@ -8,7 +8,7 @@ public class CharacterTargetEffect : ItemUseEffect
 {
     [Header("합 옵션")]
     public float recoveryHp;
-    public float recoveryStamina;
+    public int recoveryStamina;
     public int armorIncrement;
     
     [Header("곱 옵션 - 합 연산")]
